replace the old version with the refactored ones
